chore: move group schemas to separate module
